Either a rom provided on the command line is run, or a default rom.
Roms to be included with CHIP-8 are moved to src/main/resources.
These roms can be named on the command line to be run.
They are bundled with the distribution files found
  at build/distributions/fortville-0.0.1.{tar,zip}.

Any rom provided on the command line (with any pathname) will be run.
Note that a builtin rom has precedence to run over one provided on
  the command line if they have the exact same filename.

Any rom can be run from '$ gradle run' by changing the 'args' value
  in the 'run' task in the file build.gradle.